Replace InventoryView used for Packet107SetCreativeSlot with the player's activeContainer. Fixes BUKKIT-4715
The custom CraftInventoryView created before had the player's own inventory passed in as the viewing (top) inventory, when it should be the CraftInventoryCrafting, crafting matrix. Even though creative players cannot access their own crafting matrix, it IS still there, as evidenced by the slot numbers passed in from the client. Using the player's own inventory threw off the InventoryView.convertSlot() method when attempting to convert the raw slot to the inventory slot, as the top 3 rows appeared to be within the top inventory instead of the bottom inventory. This resulted in InventoryClickEvent.getSlot() returning an invalid slot number for creative players on the hotbar.

The InventoryView from the player's activeContainer is the exact same one used for survival mode players, therefore avoiding code redundancy.